requirements.txt
# AI Context Vault – Dependencies
anthropic>=0.40.0
azure-search-documents>=11.4.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
